refactor: remove component index files
