+++ app/layout.tsx
@@ -1,7 +1,7 @@
 'use client';
 import './globals.css';
 import type { ReactNode } from 'react';
-import Header from '../components/Header';
+import Header from '../components/Header/Header';
 import { I18nProvider } from '../lib/i18n';
 
 export default function RootLayout({ children }: { children: ReactNode }) {
