+++ app/my-day/page.tsx
@@ -1,5 +1,5 @@
 'use client';
-import Board from '../../components/Board';
+import Board from '../../components/Board/Board';
 
 export default function MyDayPage() {
   return (
