+++ app/my-tasks/page.tsx
@@ -1,8 +1,11 @@
 import dynamic from 'next/dynamic';
 
-const TasksView = dynamic(() => import('../../components/TasksView'), {
-  ssr: false,
-});
+const TasksView = dynamic(
+  () => import('../../components/TasksView/TasksView'),
+  {
+    ssr: false,
+  }
+);
 
 export default function MyTasksPage() {
   return <TasksView />;
